feat: Implement auto-login and auto-save
Add functionality to automatically log in users based on localStorage data and to save user progress automatically. This streamlines the user experience by persisting game state and removing the need for manual re-login.

Also, remove the explicit "reload" functionality from the MainMenu as it's no longer necessary with the new auto-login mechanism. Add a logout button to the MainMenu for explicit session termination.